use stream for logs to avoid separate logger file
